bot/requirements.txt
nonebot2[fastapi]>=2.3.0
nonebot-adapter-onebot>=2.4.0
nonebot-plugin-apscheduler>=0.4.0
httpx>=0.27.0
redis>=5.0.0
